Use Json Schema to validate video plugin data
