+++ package.json
@@ -29,6 +29,9 @@
     "@types/lodash": "^4.14.149",
     "@types/react-beautiful-dnd": "^13.1.2",
     "@types/react-pdf": "^5.0.4",
+    "ajv": "^8.10.0",
+    "ajv-errors": "^3.0.0",
+    "ajv-formats": "^2.1.1",
     "antd": "^4.9.1",
     "bodybuilder": "^2.2.25",
     "codemirror": "^5.58.2",
