+++ src/shared/containers/VideoPluginContainer/schema.json
@@ -1,0 +1,39 @@
+{
+  "$schema": "http://json-schema.org/draft-07/schema#",
+  "$id": "https://bbp.epfl.ch/nexus/plugins/video/schema/video.schema.json",
+  "type": "object",
+  "properties": {
+    "video": {
+      "oneOf": [
+        { "$ref": "#/$defs/video" },
+        { "type": "array", "items": { "$ref": "#/$defs/video" } }
+      ]
+    }
+  },
+  "required": ["video"],
+
+  "$defs": {
+    "video": {
+      "type": "object",
+      "properties": {
+        "name": {
+          "type": "string"
+        },
+        "description": {
+          "type": "string"
+        },
+        "duration": {
+          "type": "string"
+        },
+        "uploadDate": {
+          "type": "string",
+          "format": "date-time"
+        },
+        "embedUrl": {
+          "type": "string"
+        }
+      },
+      "required": ["name", "embedUrl"]
+    }
+  }
+}
